backend/pyproject.toml
[project]
name = "mycode"
version = "0.1.0"
description = "TODO"
authors = [{ name = "Olof Harrysson", email = "olof@tandemfuture.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.1.1",
    "dspy>=3.0.3",
    "fastapi>=0.116.1",
    "joblib>=1.4.2",
    "jsonargparse>=4.37.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
extend-select = [
  "UP", # pyupgrade
  # "I",  # isort
]

# Ignore unused variables
ignore = ["F841"]
